+++ src/Intermediate/Rxmxnx.PInvoke.Common.Intermediate/ValPtr.cs
@@ -7,8 +7,7 @@ namespace Rxmxnx.PInvoke;
 [Serializable]
 [StructLayout(LayoutKind.Sequential)]
 [SuppressMessage(SuppressMessageConstants.CSharpSquid, SuppressMessageConstants.CheckIdS6640)]
-public readonly unsafe struct ValPtr<T> : IWrapper<IntPtr>, IEquatable<ValPtr<T>>, IComparable, IComparable<ValPtr<T>>,
-	ISpanFormattable, ISerializable
+public readonly unsafe partial struct ValPtr<T> : IWrapper<IntPtr>, IEquatable<ValPtr<T>>, ISerializable
 {
 #pragma warning disable CS8500
 	/// <summary>
@@ -62,11 +61,6 @@ public readonly unsafe struct ValPtr<T> : IWrapper<IntPtr>, IEquatable<ValPtr<T>
 		=> ValidationUtilities.ThrowIfInvalidSerialization(info, this._value);
 
 	/// <inheritdoc/>
-	public Int32 CompareTo(Object? obj)
-		=> ValidationUtilities.ThrowIfInvalidValuePointer<T>(obj, this.Pointer, nameof(ValPtr<T>));
-	/// <inheritdoc/>
-	public Int32 CompareTo(ValPtr<T> value) => this.Pointer.CompareTo(value.Pointer);
-	/// <inheritdoc/>
 	public Boolean Equals(ValPtr<T> other) => this.Pointer == other.Pointer;
 
 	/// <inheritdoc/>
@@ -94,16 +88,6 @@ public readonly unsafe struct ValPtr<T> : IWrapper<IntPtr>, IEquatable<ValPtr<T>
 	/// </returns>
 	/// <exception cref="FormatException"><paramref name="format"/> is invalid or not supported.</exception>
 	public String ToString(String? format) => this.Pointer.ToString(format);
-	/// <inheritdoc cref="IntPtr.ToString(IFormatProvider?)"/>
-	public String ToString(IFormatProvider? formatProvider) => this.Pointer.ToString(formatProvider);
-	/// <inheritdoc/>
-	public String ToString(String? format, IFormatProvider? formatProvider)
-		=> this.Pointer.ToString(format, formatProvider);
-	/// <inheritdoc/>
-	[SuppressMessage(SuppressMessageConstants.CSharpSquid, SuppressMessageConstants.CheckIdS1006)]
-	public Boolean TryFormat(Span<Char> destination, out Int32 charsWritten, ReadOnlySpan<Char> format = default,
-		IFormatProvider? provider = default)
-		=> this.Pointer.TryFormat(destination, out charsWritten, format, provider);
 
 	/// <summary>
 	/// Retrieves an <see langword="unsafe"/> <see cref="IFixedReference{T}.IDisposable"/> instance from
@@ -216,32 +200,6 @@ public readonly unsafe struct ValPtr<T> : IWrapper<IntPtr>, IEquatable<ValPtr<T>
 	/// </summary>
 	/// <param name="pointer">The pointer to subtract the offset form.</param>
 	public static ValPtr<T> operator --(ValPtr<T> pointer) => (ValPtr<T>)(pointer.Pointer - sizeof(T));
-	/// <summary>Compares two values to determine which is less.</summary>
-	/// <param name="left">The value to compare with <paramref name="right"/>.</param>
-	/// <param name="right">The value to compare with <paramref name="left"/>.</param>
-	/// <returns><c>true</c> if <paramref name="left"/> is less than <paramref name="right"/>; otherwise, <c>false</c>.</returns>
-	public static Boolean operator <(ValPtr<T> left, ValPtr<T> right) => left.CompareTo(right) < 0;
-	/// <summary>Compares two values to determine which is less or equal.</summary>
-	/// <param name="left">The value to compare with <paramref name="right"/>.</param>
-	/// <param name="right">The value to compare with <paramref name="left"/>.</param>
-	/// <returns>
-	/// <c>true</c> if <paramref name="left"/> is less than or equal to <paramref name="right"/>; otherwise,
-	/// <c>false</c>.
-	/// </returns>
-	public static Boolean operator <=(ValPtr<T> left, ValPtr<T> right) => left.CompareTo(right) <= 0;
-	/// <summary>Compares two values to determine which is greater.</summary>
-	/// <param name="left">The value to compare with <paramref name="right"/>.</param>
-	/// <param name="right">The value to compare with <paramref name="left"/>.</param>
-	/// <returns><c>true</c> if <paramref name="left"/> is greater than <paramref name="right"/>; otherwise, <c>false</c>.</returns>
-	public static Boolean operator >(ValPtr<T> left, ValPtr<T> right) => left.CompareTo(right) > 0;
-	/// <summary>Compares two values to determine which is greater or equal.</summary>
-	/// <param name="left">The value to compare with <paramref name="right"/>.</param>
-	/// <param name="right">The value to compare with <paramref name="left"/>.</param>
-	/// <returns>
-	/// <c>true</c> if <paramref name="left"/> is greater than or equal to <paramref name="right"/>; otherwise,
-	/// <c>false</c>.
-	/// </returns>
-	public static Boolean operator >=(ValPtr<T> left, ValPtr<T> right) => left.CompareTo(right) >= 0;
 
 	/// <summary>
 	/// Adds an offset in <typeparamref name="T"/> units to the value of a pointer.
@@ -256,55 +214,5 @@ public readonly unsafe struct ValPtr<T> : IWrapper<IntPtr>, IEquatable<ValPtr<T>
 	/// <param name="offset">The offset in <typeparamref name="T"/> units to subtract.</param>
 	public static ValPtr<T> Subtract(ValPtr<T> pointer, Int32 offset)
 		=> (ValPtr<T>)(pointer.Pointer - offset * sizeof(T));
-
-	/// <inheritdoc cref="IntPtr.Parse(String)"/>
-	[ExcludeFromCodeCoverage]
-	public static ValPtr<T> Parse(String s) => (ValPtr<T>)IntPtr.Parse(s);
-	/// <inheritdoc cref="IntPtr.Parse(String, NumberStyles)"/>
-	[ExcludeFromCodeCoverage]
-	public static ValPtr<T> Parse(String s, NumberStyles style) => (ValPtr<T>)IntPtr.Parse(s, style);
-	/// <inheritdoc cref="IntPtr.Parse(String, IFormatProvider)"/>
-	[ExcludeFromCodeCoverage]
-	public static ValPtr<T> Parse(String s, IFormatProvider? provider) => (ValPtr<T>)IntPtr.Parse(s, provider);
-	/// <inheritdoc cref="IntPtr.Parse(String, NumberStyles, IFormatProvider)"/>
-	[ExcludeFromCodeCoverage]
-	public static ValPtr<T> Parse(String s, NumberStyles style, IFormatProvider? provider)
-		=> (ValPtr<T>)IntPtr.Parse(s, style, provider);
-	/// <inheritdoc cref="IntPtr.Parse(ReadOnlySpan{Char}, NumberStyles, IFormatProvider)"/>
-	[ExcludeFromCodeCoverage]
-	public static ValPtr<T> Parse(ReadOnlySpan<Char> s, NumberStyles style = NumberStyles.Integer,
-		IFormatProvider? provider = default)
-		=> (ValPtr<T>)IntPtr.Parse(s, style, provider);
-
-	/// <inheritdoc cref="IntPtr.TryParse(String?, out IntPtr)"/>
-	[ExcludeFromCodeCoverage]
-	public static Boolean TryParse([NotNullWhen(true)] String? s, out ValPtr<T> result)
-	{
-		Unsafe.SkipInit(out result);
-		return IntPtr.TryParse(s, out Unsafe.As<ValPtr<T>, IntPtr>(ref result));
-	}
-	/// <inheritdoc cref="IntPtr.TryParse(String?, NumberStyles, IFormatProvider?, out IntPtr)"/>
-	[ExcludeFromCodeCoverage]
-	public static Boolean TryParse([NotNullWhen(true)] String? s, NumberStyles style, IFormatProvider? provider,
-		out ValPtr<T> result)
-	{
-		Unsafe.SkipInit(out result);
-		return IntPtr.TryParse(s, style, provider, out Unsafe.As<ValPtr<T>, IntPtr>(ref result));
-	}
-	/// <inheritdoc cref="IntPtr.TryParse(ReadOnlySpan{Char}, out IntPtr)"/>
-	[ExcludeFromCodeCoverage]
-	public static Boolean TryParse(ReadOnlySpan<Char> s, out ValPtr<T> result)
-	{
-		Unsafe.SkipInit(out result);
-		return IntPtr.TryParse(s, out Unsafe.As<ValPtr<T>, IntPtr>(ref result));
-	}
-	/// <inheritdoc cref="IntPtr.TryParse(ReadOnlySpan{Char}, NumberStyles, IFormatProvider?, out IntPtr)"/>
-	[ExcludeFromCodeCoverage]
-	public static Boolean TryParse(ReadOnlySpan<Char> s, NumberStyles style, IFormatProvider? provider,
-		out ValPtr<T> result)
-	{
-		Unsafe.SkipInit(out result);
-		return IntPtr.TryParse(s, style, provider, out Unsafe.As<ValPtr<T>, IntPtr>(ref result));
-	}
 #pragma warning restore CS8500
 }
